Move test functions back to main
Simplifies things. The functions will be deleted after categroy
definitions is populated so the clutter should not be an issue.
